[ADD] better AbstractCSVLogger - Add new BlockchainReorgLogger that can work with any other PoW blockchain without the need for any special events - Improve PoW networks and nodes model - Add new factory methods to BitcoinBlock - Add genesis block to all consensus algorithms configs - Add new logger for uncle block counting - Add new class for PoW single parent blocks
Former-commit-id: 998be5cf95185e934f7b57205c7814abc2fa84bf [formerly 9672bc0a4ee16db668946f41dd205c31c4777ea3]
Former-commit-id: 83fdfd8619784d6834ae3ff5c49ce5e1359c1f2a

+++ src/main/java/jabs/network/networks/bitcoin/BitcoinGlobalProofOfWorkNetworkWithoutTx.java
@@ -15,7 +15,8 @@ public class BitcoinGlobalProofOfWorkNetworkWithoutTx<R extends Enum<R>> extends
     }
 
     @Override
-    public BitcoinMinerNode createSampleMiner(Simulator simulator, int nodeID, long hashPower, BitcoinBlockWithoutTx genesisBlock,
+    public BitcoinMinerNode createSampleMiner(Simulator simulator, int nodeID, double hashPower,
+                                              BitcoinBlockWithoutTx genesisBlock,
                                               ChainBasedConsensusConfig chainBasedConsensusConfig) {
         R region = this.sampleRegion();
         return new BitcoinMinerNodeWithoutTx(simulator, this, nodeID, this.sampleDownloadBandwidth(region),
