+++ src/main/java/jabs/network/stats/MinerGlobalRegionDistribution.java
@@ -3,7 +3,7 @@ package jabs.network.stats;
 public interface MinerGlobalRegionDistribution<R extends Enum<R>> {
     R sampleMinerRegion();
 
-    long sampleMinerHashPower();
+    double sampleMinerHashPower();
 
     double shareOfMinersToAllNodes();
